[IMP] test_assetsbundle: re-enable and convert test to hoot
This commit converts one of the very last qunit tests to hoot.
Fun fast: that test was actually not executed in the qunit suite,
which explains why we didn't spot it earlier, and so didn't
convert it. That test had been sneakily disabled a few years ago by
[1], presumably because it didn't pass anymore with the changes.

[1] https://github.com/odoo/odoo/pull/95500
