+++ odoo/addons/test_assetsbundle/__manifest__.py
@@ -81,6 +81,9 @@
         'test_assetsbundle.file_not_found':[
           'test_assetsbundle/static/invalid_src/xml/file_not_found.xml',
         ],
+        'web.assets_unit_tests': [
+            'test_assetsbundle/static/tests/lazy_component.test.js',
+        ],
     },
     'author': 'Odoo S.A.',
     'license': 'LGPL-3',
